requirements.txt
Django==1.5.4
Fabric==1.4.3
HireFire==0.1
PyYAML==3.10
South==0.7.6
amqp==1.0.13
anyjson==0.3.3
argparse==1.2.1
beautifulsoup4==4.3.2
billiard==2.7.3.34
boto==2.5.2
celery==3.0.24
dj-database-url==0.2.1
django-appconf==0.6
django-bootstrap-toolkit==2.14.0
django-celery==3.0.17
django-datetime-widget==0.5.3a
django-localflavor==1.0
django-notification==1.1
django-notifications-hq
django-password-reset==0.5.1
-e git://github.com/AnthonyBRoberts/django-profiles.git#egg=django-profiles
-e git://github.com/AnthonyBRoberts/django-registration-1.5.git#egg=django-registration
-e git://github.com/bitmazk/django-registration-email@3112a175a7f1fc220c6abedeb9f45094fb4ab355#egg=django_registration_email-dev
django-ses==0.4.1
django-staticfiles==1.2.1
django-storages==1.1.5
django-suit==0.2.4
django-suit-redactor==0.0.2
django-taggit==0.10a1
django-templated-email==0.4.9
django-tinymce==1.5.1
docutils==0.10
gevent==0.13.7
greenlet==0.4.0
gunicorn==0.14.6
kombu==2.5.15
lxml==2.3.5
mechanize==0.2.5
mock==1.0b1
newrelic==1.11.0.55
paramiko==1.10.1
phonenumbers==5.5b1
psycopg2==2.4.5
pycrypto==2.6
python-dateutil==2.1
pytz==2013.7
redis==2.7.1
simplejson==2.2.1
six==1.4.1
ssh==1.8.0
unittest2==0.5.1
wsgiref==0.1.2
